Small updates and reorganization of new integrator/propagator organization
