+++ include/tudat/simulation/estimation_setup/createNumericalSimulator.h
@@ -80,113 +80,6 @@ std::shared_ptr< propagators::DynamicsSimulator< StateScalarType, TimeType > > c
     }
 }
 
-////! Function to create single-arc variational equations solver object
-///*!
-// *  Function to create single-arc variational equations solver object
-// *  \param bodies Map of bodies (with names) of all bodies in integration.
-// *  \param integratorSettings Settings for numerical integrator of combined propagation of variational equations
-// *  and equations of motion.
-// *  \param propagatorSettings Settings for propagation of equations of motion.
-// *  \param parametersToEstimate Object containing all parameters that are to be estimated and their current
-// *  settings and values.
-// *  \param integrateDynamicalAndVariationalEquationsConcurrently Boolean defining whether variational and dynamical
-// *  equations are to be propagated concurrently (if true) or sequentially (of false)
-// *  \param variationalOnlyIntegratorSettings Settings for numerical integrator when integrating only variational
-// *  equations.
-// *  \param clearNumericalSolution Boolean to determine whether to clear the raw numerical solution member variables
-// *  (default true) after propagation and resetting of state transition interface.
-// *  \param integrateEquationsOnCreation Boolean to denote whether equations should be integrated immediately at the
-// *  end of this contructor.
-// *  \return Single-arc variational equations solver object
-// */
-//template< typename StateScalarType = double, typename TimeType = double >
-//std::shared_ptr< propagators::SingleArcVariationalEquationsSolver< StateScalarType, TimeType > >
-//createSingleArcVariationalEquationsSolver(
-//        const simulation_setup::SystemOfBodies& bodies,
-//        const std::shared_ptr< propagators::PropagatorSettings< StateScalarType > > propagatorSettings,
-//        const std::shared_ptr< estimatable_parameters::EstimatableParameterSet<  StateScalarType > > parametersToEstimate,
-//        const bool integrateDynamicalAndVariationalEquationsConcurrently = 1,
-//        const bool clearNumericalSolution = 1,
-//        const bool integrateEquationsOnCreation = 1 )
-//{
-//    return std::make_shared< propagators::SingleArcVariationalEquationsSolver< StateScalarType, TimeType > >(
-//                bodies, propagatorSettings, parametersToEstimate,
-//                integrateDynamicalAndVariationalEquationsConcurrently,
-//                clearNumericalSolution, integrateEquationsOnCreation );
-//}
-
-////! Function to create multi-arc variational equations solver object
-///*!
-// *  Function to create multi-arc variational equations solver object
-// *  \param bodies Map of bodies (with names) of all bodies in integration.
-// *  \param integratorSettings Settings for numerical integrator.
-// *  \param propagatorSettings Settings for propagator.
-// *  \param parametersToEstimate Object containing all parameters that are to be estimated and their current settings and values.
-// *  \param integrateDynamicalAndVariationalEquationsConcurrently Boolean defining whether variational and dynamical
-// *  equations are to be propagated concurrently (if true) or sequentially (of false)
-// *  \param variationalOnlyIntegratorSettings Settings for numerical integrator when integrating only variational
-// *  equations.
-// *  \param clearNumericalSolution Boolean to determine whether to clear the raw numerical solution member variables
-// *  (default true) after propagation and resetting of state transition interface.
-// *  \param integrateEquationsOnCreation Boolean to denote whether equations should be integrated immediately at the
-// *  end of this contructor.
-// *  \return Multi-arc variational equations solver object
-// */
-//template< typename StateScalarType = double, typename TimeType = double >
-//std::shared_ptr< propagators::MultiArcVariationalEquationsSolver< StateScalarType, TimeType > >
-//createMultiArcVariationalEquationsSolver(
-//        const simulation_setup::SystemOfBodies& bodies,
-//        const std::shared_ptr< propagators::PropagatorSettings< StateScalarType > > propagatorSettings,
-//        const std::shared_ptr< estimatable_parameters::EstimatableParameterSet<  StateScalarType > > parametersToEstimate,
-//        const bool integrateDynamicalAndVariationalEquationsConcurrently = 1,
-//        const bool clearNumericalSolution = 1,
-//        const bool integrateEquationsOnCreation = 1 )
-//{
-//    std::vector< double > arcStartTimes = estimatable_parameters::getMultiArcStateEstimationArcStartTimes(
-//                parametersToEstimate, true );
-//    return std::make_shared< propagators::MultiArcVariationalEquationsSolver< StateScalarType, TimeType > >(
-//                bodies, propagatorSettings, parametersToEstimate, arcStartTimes,
-//                integrateDynamicalAndVariationalEquationsConcurrently,
-//                clearNumericalSolution, integrateEquationsOnCreation );
-//}
-
-
-////! Function to create hybrid-arc variational equations solver object
-///*!
-// *  Function to create hybrid-arc variational equations solver object
-// *  \param bodies Map of bodies (with names) of all bodies in integration.
-// *  \param integratorSettings Settings for numerical integrator.
-// *  \param propagatorSettings Settings for propagator.
-// *  \param parametersToEstimate Object containing all parameters that are to be estimated and their current settings and values.
-// *  \param integrateDynamicalAndVariationalEquationsConcurrently Boolean defining whether variational and dynamical
-// *  equations are to be propagated concurrently (if true) or sequentially (of false)
-// *  \param variationalOnlyIntegratorSettings Settings for numerical integrator when integrating only variational
-// *  equations.
-// *  \param clearNumericalSolution Boolean to determine whether to clear the raw numerical solution member variables
-// *  (default true) after propagation and resetting of state transition interface.
-// *  \param integrateEquationsOnCreation Boolean to denote whether equations should be integrated immediately at the
-// *  end of this contructor.
-// *  \return Hybrid-arc variational equations solver object
-// */
-//template< typename StateScalarType = double, typename TimeType = double >
-//std::shared_ptr< propagators::HybridArcVariationalEquationsSolver< StateScalarType, TimeType > >
-//createHybridArcVariationalEquationsSolver(
-//        const simulation_setup::SystemOfBodies& bodies,
-//        const std::shared_ptr< propagators::PropagatorSettings< StateScalarType > > propagatorSettings,
-//        const std::shared_ptr< estimatable_parameters::EstimatableParameterSet<  StateScalarType > > parametersToEstimate,
-//        const bool integrateDynamicalAndVariationalEquationsConcurrently = 1,
-//        const bool clearNumericalSolution = 1,
-//        const bool integrateEquationsOnCreation = 1 )
-//{
-//    std::vector< double > arcStartTimes = estimatable_parameters::getMultiArcStateEstimationArcStartTimes(
-//                parametersToEstimate, false );
-
-//        return std::make_shared< propagators::HybridArcVariationalEquationsSolver< StateScalarType, TimeType > >(
-//                    bodies, propagatorSettings, parametersToEstimate, arcStartTimes,
-//                    integrateDynamicalAndVariationalEquationsConcurrently,
-//                    clearNumericalSolution, integrateEquationsOnCreation );
-//}
-
 //! Function to create variational equations solver object
 /*!
 *  Function to create variational equations solver object
