+++ src/simulation/propagation_setup/CMakeLists.txt
@@ -12,7 +12,6 @@
 set(propagation_HEADERS
         createAccelerationModels.h
         dynamicsSimulator.h
-#        propagationCR3BPFullProblem.h
         createTorqueModel.h
         createStateDerivativeModel.h
         createEnvironmentUpdater.h
@@ -22,20 +21,17 @@ set(propagation_HEADERS
         createMassRateModels.h
         propagationTermination.h
         propagationSettings.h
+        propagationPrintSettings.h
+        propagationProcessingSettings.h
         accelerationSettings.h
         propagationOutputSettings.h
         setNumericallyIntegratedStates.h
         environmentUpdater.h
-#        propagationLambertTargeterFullProblem.h
-#<<<<<<< HEAD
-#        propagationPatchedConicFullProblem.h
-#        propagationLowThrustProblem.h
         )
 
 # Add header files.
 set(propagation_SOURCES
         createStateDerivativeModel.cpp
-#        propagationLambertTargeterFullProblem.cpp
         propagationSettings.cpp
         setNumericallyIntegratedStates.cpp
         dynamicsSimulator.cpp
@@ -44,13 +40,10 @@ set(propagation_SOURCES
         createMassRateModels.cpp
         createAccelerationModels.cpp
         createEnvironmentUpdater.cpp
-#        propagationCR3BPFullProblem.cpp
         propagationTermination.cpp
         propagationOutput.cpp
         environmentUpdater.cpp
-#<<<<<<< HEAD
-#        propagationPatchedConicFullProblem.cpp
-#        propagationLowThrustProblem.cpp
+
         )
 
 # Add library.
