+++ tests/src/astro/orbit_determination/unitTestHybridArcStateEstimation.cpp
@@ -9,8 +9,8 @@
  *
  */
 
-//#define BOOST_TEST_DYN_LINK
-//#define BOOST_TEST_MAIN
+#define BOOST_TEST_DYN_LINK
+#define BOOST_TEST_MAIN
 
 #include <string>
 #include <thread>
@@ -28,11 +28,11 @@
 #include "tudat/simulation/environment_setup/createGroundStations.h"
 
 
-//namespace tudat
-//{
-//namespace unit_tests
-//{
-//BOOST_AUTO_TEST_SUITE( test_hybrid_arc_state_estimation )
+namespace tudat
+{
+namespace unit_tests
+{
+BOOST_AUTO_TEST_SUITE( test_hybrid_arc_state_estimation )
 
 //Using declarations.
 using namespace tudat;
@@ -360,43 +360,42 @@ Eigen::VectorXd  executeParameterEstimation(
 }
 
 
-//BOOST_AUTO_TEST_CASE( test_HybridArcStateEstimation )
-int main( )
+BOOST_AUTO_TEST_CASE( test_HybridArcStateEstimation )
 {
     // Perform estimation
     Eigen::VectorXd parameterError = executeParameterEstimation< double, double, double >( );
     int numberOfEstimatedArcs = ( parameterError.rows( ) - 8 ) / 6;
 
-//    std::cout<<std::endl<<std::endl<<"Final error: "<<parameterError.transpose( )<<std::endl;
-//    // Test error range: 5 m in-plane position and 2 micron/s in-plane velocity for Mars
-//    for( unsigned int j = 0; j < 2; j++ )
-//    {
-//        BOOST_CHECK_SMALL( std::fabs( parameterError( j ) ), 5.0 );
-//        BOOST_CHECK_SMALL( std::fabs( parameterError( j + 3 ) ), 2.0E-6  );
-//    }
+    std::cout<<std::endl<<std::endl<<"Final error: "<<parameterError.transpose( )<<std::endl;
+    // Test error range: 5 m in-plane position and 2 micron/s in-plane velocity for Mars
+    for( unsigned int j = 0; j < 2; j++ )
+    {
+        BOOST_CHECK_SMALL( std::fabs( parameterError( j ) ), 5.0 );
+        BOOST_CHECK_SMALL( std::fabs( parameterError( j + 3 ) ), 2.0E-6  );
+    }
 
-//    // Test error range: 1000 m in-plane position and 0.5 mm/s in-plane velocity for Mars (poor values due to short arc)
-//    BOOST_CHECK_SMALL( std::fabs( parameterError( 2 ) ), 1000.0 );
-//    BOOST_CHECK_SMALL( std::fabs( parameterError( 5 ) ), 0.5E-3  );
+    // Test error range: 1000 m in-plane position and 0.5 mm/s in-plane velocity for Mars (poor values due to short arc)
+    BOOST_CHECK_SMALL( std::fabs( parameterError( 2 ) ), 1000.0 );
+    BOOST_CHECK_SMALL( std::fabs( parameterError( 5 ) ), 0.5E-3  );
 
-//    // Test error range: 0.1 m position and 50 micron/s velocity for orbiter
-//    for( int i = 0; i < numberOfEstimatedArcs; i++ )
-//    {
-//        for( unsigned int j = 0; j < 3; j++ )
-//        {
-//            BOOST_CHECK_SMALL( std::fabs( parameterError( ( i + 1 )* 6 + j ) ), 1E-1 );
-//            BOOST_CHECK_SMALL( std::fabs( parameterError( ( i + 1 ) * 6 + j + 3 ) ), 5.0E-5  );
-//        }
-//    }
+    // Test error range: 0.1 m position and 50 micron/s velocity for orbiter
+    for( int i = 0; i < numberOfEstimatedArcs; i++ )
+    {
+        for( unsigned int j = 0; j < 3; j++ )
+        {
+            BOOST_CHECK_SMALL( std::fabs( parameterError( ( i + 1 )* 6 + j ) ), 1E-1 );
+            BOOST_CHECK_SMALL( std::fabs( parameterError( ( i + 1 ) * 6 + j + 3 ) ), 5.0E-5  );
+        }
+    }
 
-//    // Test errors for gravitational parameters
-//    BOOST_CHECK_SMALL( std::fabs( parameterError( parameterError.rows( ) - 2 ) ), 1.5E11 );
-//    BOOST_CHECK_SMALL( std::fabs( parameterError( parameterError.rows( ) - 1 ) ), 1.0E6 );
+    // Test errors for gravitational parameters
+    BOOST_CHECK_SMALL( std::fabs( parameterError( parameterError.rows( ) - 2 ) ), 1.5E11 );
+    BOOST_CHECK_SMALL( std::fabs( parameterError( parameterError.rows( ) - 1 ) ), 1.0E6 );
 
 }
 
-//BOOST_AUTO_TEST_SUITE_END( )
+BOOST_AUTO_TEST_SUITE_END( )
 
-//}
+}
 
-//}
+}
